Web UI v0.1
First version of the site. It has been hardcoded, as it's said in the
requirements there is no need for DB or anything. From categories, only
bread has data and from recipes, only pasta --> bolognesa. Discounts
are not implemented yet.
Also the shopping cart is based on paypal, I've talked to Sampo and he
said it's fine to use 3rd part apps as long as it is stated in the
report.